Updating code to combine SearchSettings and Search Impl as well as efficiency improvements.
