+++ jadx-gui/src/main/java/jadx/gui/utils/search/TextSearchIndex.java
@@ -99,13 +99,7 @@ public class TextSearchIndex {
 		LOG.debug("Building search, ignoreCase: {}, useRegex: {}", ignoreCase, useRegex);
 		Flowable<JNode> result = Flowable.empty();
 
-		SearchSettings searchSettings = new SearchSettings();
-		searchSettings.setSearchString(text);
-		searchSettings.setIgnoreCase(options.contains(IGNORE_CASE));
-		searchSettings.setRegex(options.contains(USE_REGEX));
-		if (!searchSettings.preCompile()) {
-			return result;
-		}
+		SearchSettings searchSettings = new SearchSettings(text, options.contains(IGNORE_CASE), options.contains(USE_REGEX));
 
 		if (options.contains(CLASS)) {
 			result = Flowable.concat(result, clsNamesIndex.search(searchSettings));
@@ -154,7 +148,7 @@ public class TextSearchIndex {
 
 	private int searchNext(FlowableEmitter<CodeNode> emitter, JavaNode javaClass, String code, final SearchSettings searchSettings) {
 		int pos;
-		pos = SearchImpl.find(code, searchSettings);
+		pos = searchSettings.find(code, searchSettings);
 		if (pos == -1) {
 			return -1;
 		}
